refactor(cli): extract CLI report rendering into report module
Move the pure CLI report-rendering seam out of the main.rs binary crate
into a new `s11::report` library module, continuing the main.rs
decomposition started in #729 and #739.

The cluster is a family of pure functions that turn search results, LLM
timing breakdowns, unsupported-mnemonic ledgers and equivalence outcomes
into the exact lines the CLI prints:

  - fmt_bytes / fmt_dur
  - format_llm_timings
  - format_unsupported_mnemonic_ledger
  - format_search_statistics
  - EquivReport / build_equiv_report

They had zero coupling to any main-local struct and were already written
as pure `Vec<String>`-returning seams for testability, but living in the
binary crate meant they were unreachable from integration tests. The
binary keeps the thin `print_*` wrappers that loop over these lines and
`println!`, and `run_equiv`, so stdout stays out of the library
(matching the capstone_bridge precedent).

main.rs: 6939 -> 6446 lines. No behavior change: the moved code is
byte-identical and the relocated unit tests pin the exact output.

Adds tests/integration/report.rs, which exercises the seam through
`s11::report::*` from outside the crate — coverage that was impossible
before this change.
