+++ tests/integration/mod.rs
@@ -5,3 +5,4 @@ mod docs_capability;
 mod equiv_test;
 mod live_out_cli_test;
 mod opt_test;
+mod report;
